Restructured layout-based components
