+++ app/src/components/JustPlaces/RatingComponent.css
