+++ app/src/components/JustPlaces/RatingComponent.tsx
