+++ app/src/components/ui/SearchResult.css
